make the new collection class a separate file for transition
