+++ src/easyscience/base_classes/__init__.py
@@ -1,8 +1,9 @@
 from .based_base import BasedBase
+from .collection_base_easylist import CollectionBase as CollectionBaseEasyList
 from .collection_base import CollectionBase
 from .easy_list import EasyList
 from .model_base import ModelBase
 from .new_base import NewBase
 from .obj_base import ObjBase
 
-__all__ = [BasedBase, CollectionBase, ObjBase, ModelBase, NewBase, EasyList]
+__all__ = [BasedBase, CollectionBase, CollectionBaseEasyList, ObjBase, ModelBase, NewBase, EasyList]
